fix(core): resolve core→features coupling violations
Move shared env result types (EnvLogsOptions/Result, EnvStartResult,
EnvStopResult) to core/runtime/env-types.ts so RuntimeAdapter can
reference them without importing from features/. Move resolveLiferayConfig
to core/config/liferay-connection-config.ts with an explicit
scopeAliasDefault parameter; the features/ wrapper fills in
DEFAULT_OAUTH_SCOPE_ALIASES_STRING. Move LdevNativeRuntimeAdapter and
runtime-adapter-factory to features/env/ since they call feature
functions — BladeWorkspaceRuntimeAdapter stays in core/ as it has no
feature deps. Remove all five eslint-disable no-restricted-imports
silences from src/core/.
