+++ src/core/http/config.ts
@@ -1,5 +1,6 @@
-// TODO: pre-existing core→features coupling; this re-export barrel should move to features/ or the types should move to core/ (see architecture audit)
-/* eslint-disable no-restricted-imports */
-export type {ResolvedLiferayConfigInput} from '../../features/liferay/liferay-connection-config.js';
-export {resolveLiferayConfig} from '../../features/liferay/liferay-connection-config.js';
-/* eslint-enable no-restricted-imports */
+/**
+ * Re-exports Liferay connection config utilities from core/config.
+ * Types and resolution logic live in core/config/liferay-connection-config.ts.
+ */
+export type {ResolvedLiferayConfigInput} from '../config/liferay-connection-config.js';
+export {resolveLiferayConfig} from '../config/liferay-connection-config.js';
