+++ src/core/runtime/ldev-native-runtime-adapter.ts
@@ -1,32 +1,5 @@
-import type {AppConfig} from '../config/load-config.js';
-import type {RuntimeAdapter, RuntimeStartOptions, RuntimeStopOptions} from './runtime-adapter.js';
-// TODO: pre-existing core→features coupling; env functions and result types should be defined in core/runtime/ (see architecture audit)
-/* eslint-disable no-restricted-imports */
-import {runEnvLogs, type EnvLogsOptions, type EnvLogsResult} from '../../features/env/env-logs.js';
-import {runEnvStart, type EnvStartResult} from '../../features/env/env-start.js';
-import {runEnvStatus} from '../../features/env/env-status.js';
-import {runEnvStop, type EnvStopResult} from '../../features/env/env-stop.js';
-import type {EnvStatusReport} from '../../features/env/env-health.js';
-/* eslint-enable no-restricted-imports */
-
-export class LdevNativeRuntimeAdapter implements RuntimeAdapter {
-  readonly kind = 'ldev-native';
-
-  constructor(private readonly config: AppConfig) {}
-
-  start(options?: RuntimeStartOptions): Promise<EnvStartResult> {
-    return runEnvStart(this.config, options);
-  }
-
-  stop(options?: RuntimeStopOptions): Promise<EnvStopResult> {
-    return runEnvStop(this.config, options);
-  }
-
-  status(options?: {processEnv?: NodeJS.ProcessEnv}): Promise<EnvStatusReport> {
-    return runEnvStatus(this.config, options);
-  }
-
-  logs(options?: EnvLogsOptions): Promise<EnvLogsResult> {
-    return runEnvLogs(this.config, options);
-  }
-}
+/**
+ * @deprecated Moved to features/env/ldev-native-runtime-adapter.ts
+ * This file is kept as a tombstone. Import from features/env/runtime-adapter-factory.ts instead.
+ */
+export {};
