+++ src/features/liferay/liferay-connection-config.ts
@@ -1,54 +1,20 @@
-import {resolveRuntimeNumberConfig, resolveRuntimeStringConfig} from '../../core/config/liferay-runtime-config.js';
+import {resolveLiferayConfig as resolveLiferayConfigCore} from '../../core/config/liferay-connection-config.js';
 import {DEFAULT_OAUTH_SCOPE_ALIASES_STRING} from '../oauth/oauth-scope-aliases.js';
 
-export type ResolvedLiferayConfigInput = {
+export type {ResolvedLiferayConfigInput} from '../../core/config/liferay-connection-config.js';
+
+/**
+ * Resolves the Liferay connection config, applying the default OAuth scope aliases.
+ * This wrapper exists so callers in features/ don't need to pass the scope default explicitly;
+ * the full core function is in core/config/liferay-connection-config.ts.
+ */
+export function resolveLiferayConfig(input: {
   processEnv: NodeJS.ProcessEnv;
   dockerEnv: Record<string, string>;
   localProfile: Record<string, string>;
-};
-
-export function resolveLiferayConfig(input: ResolvedLiferayConfigInput) {
-  const {processEnv, dockerEnv, localProfile} = input;
-
-  const bindIpValue = dockerEnv.BIND_IP;
-  const rawBindIp = typeof bindIpValue === 'string' ? bindIpValue.trim() : '';
-  const hasSpecificBindIp = rawBindIp !== '' && !['0.0.0.0', '127.0.0.1', 'localhost'].includes(rawBindIp);
-  const bindIp = hasSpecificBindIp ? rawBindIp : 'localhost';
-  const httpPortValue: unknown = dockerEnv.LIFERAY_HTTP_PORT;
-  const httpPort = typeof httpPortValue === 'string' && httpPortValue.trim() !== '' ? httpPortValue.trim() : '8080';
-  const bindIpUrl = hasSpecificBindIp ? `http://${bindIp}:${httpPort}` : undefined;
-  const fallbackUrl = `http://${bindIp}:${httpPort}`;
-
-  return {
-    url: resolveRuntimeStringConfig({
-      envValue: processEnv.LIFERAY_CLI_URL,
-      localProfileValue: localProfile['liferay.url'],
-      dockerEnvValue: dockerEnv.LIFERAY_CLI_URL,
-      fallbackValue: bindIpUrl || fallbackUrl,
-    }),
-    oauth2ClientId: resolveRuntimeStringConfig({
-      envValue: processEnv.LIFERAY_CLI_OAUTH2_CLIENT_ID,
-      localProfileValue: localProfile['liferay.oauth2.clientId'],
-      dockerEnvValue: dockerEnv.LIFERAY_CLI_OAUTH2_CLIENT_ID,
-      fallbackValue: '',
-    }),
-    oauth2ClientSecret: resolveRuntimeStringConfig({
-      envValue: processEnv.LIFERAY_CLI_OAUTH2_CLIENT_SECRET,
-      localProfileValue: localProfile['liferay.oauth2.clientSecret'],
-      dockerEnvValue: dockerEnv.LIFERAY_CLI_OAUTH2_CLIENT_SECRET,
-      fallbackValue: '',
-    }),
-    scopeAliases: resolveRuntimeStringConfig({
-      envValue: processEnv.LIFERAY_CLI_OAUTH2_SCOPE_ALIASES,
-      localProfileValue: localProfile['liferay.oauth2.scopeAliases'],
-      dockerEnvValue: dockerEnv.LIFERAY_CLI_OAUTH2_SCOPE_ALIASES,
-      fallbackValue: DEFAULT_OAUTH_SCOPE_ALIASES_STRING,
-    }),
-    timeoutSeconds: resolveRuntimeNumberConfig({
-      envValue: processEnv.LIFERAY_CLI_HTTP_TIMEOUT_SECONDS,
-      localProfileValue: localProfile['liferay.oauth2.timeoutSeconds'],
-      dockerEnvValue: dockerEnv.LIFERAY_CLI_HTTP_TIMEOUT_SECONDS,
-      fallbackValue: 30,
-    }),
-  };
+}) {
+  return resolveLiferayConfigCore({
+    ...input,
+    scopeAliasDefault: DEFAULT_OAUTH_SCOPE_ALIASES_STRING,
+  });
 }
